Move upgrade flow into its own package
